backend/requirements.txt
fastapi
pydantic
uvicorn
aiofiles
python-multipart
SQLAlchemy
asyncpg
pytz
passlib
pydantic[email]
bcrypt
requests
python-dotenv